+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug in ScenarioDbManager.update_scenario_output_tables_in_db
 - ### Added
 - Added DataManager.set_parameters()
+- Added print of num rows and columns inserting in ScenarioDbManager.update_scenario_output_tables_in_db
+- ScenarioDbManager.insert_scenarios_from_zip
 
 ## [0.5.4.0] - 2022-01-11
 ### Changed